restore part of the extraction logic
