+++ lib/mindee.rb
@@ -7,6 +7,10 @@ module Mindee
   module Errors
   end
 
+  # Custom extraction module
+  module Extraction
+  end
+
   # Mindee internal http module.
   module HTTP
   end
