+++ lib/mindee/extraction.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative 'extraction/multi_receipts_extractor'
