+++ spec/extraction/multi_receipts_extractor_spec.rb
@@ -2,6 +2,7 @@
 
 require 'mindee/product'
 require 'mindee/input/sources'
+require 'mindee/extraction'
 require_relative '../data'
 
 describe 'multi-receipts extraction' do
@@ -37,7 +38,7 @@ describe 'multi-receipts extraction' do
       input_sample = Mindee::Input::Source::PathInputSource.new(multi_receipts_single_page_path)
       response = load_json(multi_receipts_single_page_json_path, 'complete.json')
       doc = Mindee::Product::MultiReceiptsDetector::MultiReceiptsDetectorV1.new(response['document']['inference'])
-      extracted_receipts = Mindee::Image::ImageExtractor.extract_receipts(input_sample, doc)
+      extracted_receipts = Mindee::Extraction.extract_receipts(input_sample, doc)
 
       expect(extracted_receipts.size).to eq(6)
 
@@ -85,7 +86,7 @@ describe 'multi-receipts extraction' do
       input_sample = Mindee::Input::Source::PathInputSource.new(multi_receipts_multi_page_path)
       response = load_json(multi_receipts_multi_page_json_path, 'multipage_sample.json')
       doc = Mindee::Product::MultiReceiptsDetector::MultiReceiptsDetectorV1.new(response['document']['inference'])
-      extracted_receipts = Mindee::Image::ImageExtractor.extract_receipts(input_sample, doc)
+      extracted_receipts = Mindee::Extraction.extract_receipts(input_sample, doc)
 
       expect(extracted_receipts.size).to eq(5)
 
@@ -124,7 +125,7 @@ describe 'multi-receipts extraction' do
   context 'when no receipts are found in inference' do
     it 'raises a MindeeInputError' do
       expect do
-        Mindee::Image::ImageExtractor.extract_receipts(empty_input_source, empty_inference)
+        Mindee::Extraction.extract_receipts(empty_input_source, empty_inference)
       end.to raise_error(Mindee::Errors::MindeeInputError,
                          'No possible receipts candidates found for Multi-Receipts extraction.')
     end
@@ -132,8 +133,8 @@ describe 'multi-receipts extraction' do
 
   context 'when input source has no pages' do
     it 'returns an empty array' do
-      extracted_receipts = Mindee::Image::ImageExtractor.extract_receipts(empty_input_source,
-                                                                          valid_inference_with_no_receipts)
+      extracted_receipts = Mindee::Extraction.extract_receipts(empty_input_source,
+                                                               valid_inference_with_no_receipts)
       expect(extracted_receipts).to eq([])
     end
   end
